requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.35
PyJWT==2.9.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-jose[cryptography]==3.3.0
httpx==0.27.2
pandas==2.2.2
openpyxl==3.1.5
psutil==5.9.8